requirements.txt
PyQt6
brainglobe-atlasapi
pynrrd
numpy
pandas
opencv-python
xmltodict
scipy


